feat: move web folder to lib and static to assets
