+++ apps/juicebox_web/assets/js/containers/app-content.js
