+++ apps/juicebox_web/assets/js/containers/root.dev.js
