+++ apps/juicebox_web/assets/js/sockets/socket-middleware.js
