+++ apps/juicebox_web/assets/js/stores/main-store.js
