+++ apps/juicebox_web/assets/js/stores/reducers.js
