+++ apps/juicebox_web/assets/js/videos/actions.js
