+++ apps/juicebox_web/assets/js/videos/reducers.js
