+++ apps/juicebox_web/lib/web/web.ex
@@ -40,7 +40,7 @@ defmodule JuiceboxWeb.Web do
 
   def view do
     quote do
-      use Phoenix.View, root: "web/templates"
+      use Phoenix.View, root: "lib/web/templates"
 
       # Import convenience functions from controllers
       import Phoenix.Controller, only: [get_csrf_token: 0, get_flash: 2, view_module: 1]
